requirements.txt
requests>=2.31.0
Pillow>=10.4.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
cloudinary>=1.40.0
rich>=13.7.0
